chore(): remove unused login hooks, rename hook file to .new
