+++ ui/apps/akasha/jest.setup.js
@@ -3,7 +3,7 @@ import '../jest.setup';
 import { genLoggedInState, genPostData, genCommentData } from '@akashaorg/af-testing';
 import * as postHooks from '@akashaorg/ui-awf-hooks/lib/use-posts';
 import * as commentHooks from '@akashaorg/ui-awf-hooks/lib/use-comments';
-import * as loginHooks from '@akashaorg/ui-awf-hooks/lib/use-login';
+import * as loginHooks from 'ui/hooks/src/use-login.new.js';
 import * as mediaHooks from '@akashaorg/ui-awf-hooks/lib/utils/media-utils';
 
 jest.mock('@akashaorg/typings/ui', () => ({
